requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
openai>=1.40.0
pydantic>=2.7.0
reportlab>=4.2.0
python-multipart>=0.0.9